add comments and some mroe renames
