+++ runtime/src/iree-amd-aie/aie_runtime/CMakeLists.txt
@@ -538,12 +538,12 @@ iree_cc_library(
   NAME
     iree_aie_runtime_static
   HDRS
-    iree_aie_cdo_emitter.h
+    iree_aie_configure.h
     iree_aie_router.h
     iree_aie_runtime.h
     xaie_hwcfg.h
   SRCS
-    iree_aie_cdo_emitter.cc
+    iree_aie_configure.cc
     iree_aie_router.cc
     iree_aie_runtime.cc
     mlir_aie_legacy.cc
