+++ runtime/src/iree-amd-aie/aie_runtime/mlir_aie_legacy.cc
@@ -3,6 +3,7 @@
 // Licensed under the Apache License v2.0 with LLVM Exceptions. See
 // https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: # Apache-2.0 WITH LLVM-exception
+
 //===----------------------------------------------------------------------===//
 //
 // This file holds legacy (adapted from mlir-aie) device model data for xcvc1902
