+++ runtime/src/iree-amd-aie/aie_runtime/test/CMakeLists.txt
@@ -46,3 +46,5 @@ target_include_directories(
   iree-amd-aie_aie_runtime_test_DeviceModelTestCppTest
   PRIVATE "${IREE_MLIR_AIE_SOURCE_DIR}/include"
 )
+
+add_subdirectory(cdo)
